requirements.txt
numpy==2.4.1
pdf2image==1.17.0
pillow==12.1.0
